reorganized the type declarations and included a new index.ts to export Desolver
